@@ -36,7 +36,7 @@ class StatisticClear extends Command
     /**
      * Create a new command instance.
      *
-     * @param \App\Repositories\WorldRepository $worldRepository
+     * @param  \App\Repositories\WorldRepository  $worldRepository
      */
     public function __construct(WorldRepository $worldRepository)
     {
