@@ -12,8 +12,8 @@ class ForceJsonResponse
     /**
      * Handle an incoming request.
      *
-     * @param Request  $request
-     * @param \Closure $next
+     * @param  Request   $request
+     * @param  \Closure  $next
      * @return mixed
      */
     public function handle(Request $request, Closure $next)
