@@ -4,7 +4,6 @@ namespace App\Providers;
 
 
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
-use Illuminate\Support\Facades\Auth;
 
 
 class AuthServiceProvider extends ServiceProvider
@@ -15,9 +14,8 @@ class AuthServiceProvider extends ServiceProvider
      * @var array
      */
     protected $policies = [
-        // 'App\Model' => 'App\Policies\ModelPolicy',
+        // 'App\Models\Model' => 'App\Policies\ModelPolicy',
     ];
-
 
     /**
      * Register any authentication / authorization services.
@@ -28,8 +26,5 @@ class AuthServiceProvider extends ServiceProvider
     {
         $this->registerPolicies();
 
-        Auth::provider('custom', function ($app, array $config) {
-            return new CustomUserProvider();
-        });
     }
 }
